+++ selfdrive/carrot/cluster/cluster_ui.py
@@ -5,7 +5,6 @@ from cluster_models import (
     ClusterUiState,
     DetectedVehicle,
     LaneMarking,
-    LeadTrajectoryPoint,
     ModelPathPoint,
     ModelRiskPoint,
     RadarPoint,
@@ -39,7 +38,6 @@ __all__ = [
     "ClusterUiState",
     "DetectedVehicle",
     "LaneMarking",
-    "LeadTrajectoryPoint",
     "MeshStrip",
     "ModelPathPoint",
     "ModelRiskPoint",
