+++ selfdrive/carrot/cluster/main.py
@@ -1,109 +1,24 @@
 from __future__ import annotations
 
 import argparse
-from dataclasses import replace
 import gc
-import math
-import random
-import threading
+from dataclasses import replace
 import time
 from pathlib import Path
-from typing import Any
 
-from cluster_config import CLUSTER_THEME_PARAM, normalize_cluster_theme_mode
+from cluster_config import CLUSTER_THEME_PARAM, DESIGN_HEIGHT, DESIGN_WIDTH, normalize_cluster_theme_mode
+from cluster_gamepad import DualSenseSimulator
 from cluster_live import OpenpilotLiveSource
+from cluster_models import RouteOverlay, SimulatorInput
+from cluster_profile import GcProfileHook, ProfileReporter, freeze_gc_after_init
+from cluster_renderer import ClusterUiRenderer
 from cluster_route_replay import RouteReplaySource
+from cluster_simulator import ClusterSimulator, RandomInputSource
 from cluster_usb_display import TuringUsbDisplay
-from cluster_ui import (
-    DESIGN_HEIGHT,
-    DESIGN_WIDTH,
-    MAX_SPEED_KPH,
-    SURROUND_MAX_PITCH_DEG,
-    SURROUND_MAX_YAW_DEG,
-    ClusterSimulator,
-    ClusterUiRenderer,
-    RandomInputSource,
-    RouteOverlay,
-    SimulatorInput,
-)
+from cluster_usb_pipeline import AsyncJpegUsbPipeline
 
 DEFAULT_FPS = 0.0
-TRIGGER_DEADZONE = 0.03
-STEERING_DEADZONE = 0.06
-VIEW_ROTATION_DEADZONE = 0.08
-GAMEPAD_WARMUP_SECONDS = 0.6
-LEFT_SIGNAL_BUTTONS = (4, 9, 13)
-RIGHT_SIGNAL_BUTTONS = (5, 10, 14)
 THEME_PARAM_POLL_SECONDS = 1.0
-
-
-class ProfileReporter:
-    def __init__(self, enabled: bool, interval_s: float) -> None:
-        self.enabled = enabled
-        self.interval_s = max(0.2, interval_s)
-        self.samples: dict[str, list[float]] = {}
-        self.last_report_time = time.perf_counter()
-        self.report_frames = 0
-
-    def add(self, name: str, milliseconds: float) -> None:
-        if not self.enabled:
-            return
-        self.samples.setdefault(name, []).append(milliseconds)
-
-    def add_elapsed(self, name: str, start_time: float) -> None:
-        if self.enabled:
-            self.add(name, (time.perf_counter() - start_time) * 1000.0)
-
-    def add_samples(self, samples: tuple[tuple[str, float], ...]) -> None:
-        if not self.enabled:
-            return
-        for name, milliseconds in samples:
-            self.add(name, milliseconds)
-
-    def frame_done(self) -> None:
-        if self.enabled:
-            self.report_frames += 1
-
-    def maybe_report(self, now: float) -> None:
-        if not self.enabled or now - self.last_report_time < self.interval_s:
-            return
-
-        elapsed = max(0.001, now - self.last_report_time)
-        print(f"PROFILE {self.report_frames} frames / {elapsed:.1f}s", flush=True)
-        ordered = sorted(
-            self.samples.items(),
-            key=lambda item: sum(item[1]) / max(1, len(item[1])),
-            reverse=True,
-        )
-        for name, values in ordered:
-            if not values:
-                continue
-            average = sum(values) / len(values)
-            print(
-                f"  {name:<42} avg={average:7.2f}ms "
-                f"max={max(values):7.2f}ms last={values[-1]:7.2f}ms n={len(values)}",
-                flush=True,
-            )
-        self.samples.clear()
-        self.report_frames = 0
-        self.last_report_time = now
-
-
-class GcProfileHook:
-    def __init__(self, profile: ProfileReporter) -> None:
-        self.profile = profile
-        self._starts: dict[int, float] = {}
-
-    def __call__(self, phase: str, info: dict[str, Any]) -> None:
-        generation = int(info.get("generation", -1))
-        if phase == "start":
-            self._starts[generation] = time.perf_counter()
-            return
-        if phase != "stop":
-            return
-        start_time = self._starts.pop(generation, None)
-        if start_time is not None:
-            self.profile.add(f"gc.gen{generation}", (time.perf_counter() - start_time) * 1000.0)
 
 
 class ClusterThemeParamReader:
@@ -125,310 +40,12 @@ class ClusterThemeParamReader:
             return "auto"
 
 
-def freeze_gc_after_init(profile: ProfileReporter) -> None:
-    freeze = getattr(gc, "freeze", None)
-    if freeze is None:
-        return
-
-    profile_stage = time.perf_counter()
-    gc.collect(2)
-    profile.add_elapsed("gc.freeze_init.collect", profile_stage)
-
-    profile_stage = time.perf_counter()
-    freeze()
-    profile.add_elapsed("gc.freeze_init.freeze", profile_stage)
-
-
-class AsyncJpegUsbPipeline:
-    def __init__(self, usb_display: TuringUsbDisplay) -> None:
-        self.usb_display = usb_display
-        self._condition = threading.Condition()
-        self._pending_rgba: tuple[bytes, int, int] | None = None
-        self._closing = False
-        self._error: BaseException | None = None
-        self._samples: list[tuple[str, float]] = []
-        self._thread = threading.Thread(target=self._run, name="cluster-usb-jpeg", daemon=True)
-
-    def start(self) -> None:
-        self._thread.start()
-
-    def submit_rgba(self, rgba: bytes, width: int, height: int) -> None:
-        self.check_error()
-        with self._condition:
-            self._pending_rgba = (rgba, width, height)
-            self._condition.notify()
-
-    def profile_samples(self) -> tuple[tuple[str, float], ...]:
-        with self._condition:
-            samples = tuple(self._samples)
-            self._samples.clear()
-        return samples
-
-    def check_error(self) -> None:
-        if self._error is not None:
-            raise RuntimeError("asynchronous USB JPEG pipeline failed") from self._error
-
-    def close(self) -> None:
-        with self._condition:
-            self._closing = True
-            self._condition.notify()
-        self._thread.join(timeout=3.0)
-
-    def _add_sample(self, name: str, start_time: float) -> None:
-        milliseconds = (time.perf_counter() - start_time) * 1000.0
-        with self._condition:
-            self._samples.append((name, milliseconds))
-
-    def _add_samples(self, samples: tuple[tuple[str, float], ...]) -> None:
-        if not samples:
-            return
-        with self._condition:
-            self._samples.extend(samples)
-
-    def _take_pending(self) -> tuple[bytes, int, int] | None:
-        with self._condition:
-            while self._pending_rgba is None and not self._closing:
-                self._condition.wait(timeout=0.1)
-            if self._pending_rgba is None:
-                return None
-            pending = self._pending_rgba
-            self._pending_rgba = None
-            return pending
-
-    def _run(self) -> None:
-        while True:
-            pending = self._take_pending()
-            if pending is None:
-                return
-            rgba, width, height = pending
-            try:
-                self.usb_display.clear_profile_samples()
-                profile_stage = time.perf_counter()
-                jpeg = self.usb_display.encode_jpeg(rgba, width, height)
-                self._add_sample("usb_async.encode_jpeg", profile_stage)
-
-                profile_stage = time.perf_counter()
-                self.usb_display.send_jpeg(jpeg)
-                self._add_sample("usb_async.send_jpeg", profile_stage)
-                self._add_samples(self.usb_display.profile_samples())
-                self.usb_display.clear_profile_samples()
-            except BaseException as exc:
-                self._error = exc
-                return
-
-
-def next_simulated_values(speed_kph: float, dt: float) -> tuple[float, float, float]:
-    accel_mps2 = 1.8 * math.sin(time.monotonic() * 0.8) + random.uniform(-0.35, 0.35)
-    next_speed = max(0.0, min(MAX_SPEED_KPH, speed_kph + accel_mps2 * dt * 3.6))
-    steering = 0.42 * math.sin(time.monotonic() * 0.55)
-    return next_speed, accel_mps2, steering
-
-
 def route_overlay_for_mode(overlay: RouteOverlay | None, mode: str) -> RouteOverlay | None:
     if overlay is None or mode == "off":
         return None
     if mode == "compact":
         return replace(overlay, data_lines=overlay.data_lines[:4])
     return overlay
-
-
-def normalize_signed_axis(axis_value: float | int) -> float:
-    value = float(axis_value)
-    if value < -1.0 or value > 1.0:
-        value = value / 32767.0 if value >= 0 else value / 32768.0
-    return max(-1.0, min(1.0, value))
-
-
-def normalize_trigger_axis(axis_value: float | int) -> float:
-    value = float(axis_value)
-    if value < -1.0 or value > 1.0:
-        if value < 0.0:
-            value = (value + 32768.0) / 65535.0
-        else:
-            value = value / 32767.0
-    elif value < 0.0:
-        value = (value + 1.0) * 0.5
-
-    if value < TRIGGER_DEADZONE:
-        return 0.0
-    return max(0.0, min(1.0, value))
-
-
-def normalize_stick(axis_value: float) -> float:
-    if abs(axis_value) < STEERING_DEADZONE:
-        return 0.0
-    return max(-1.0, min(1.0, axis_value))
-
-
-def normalize_view_axis(axis_value: float) -> float:
-    if abs(axis_value) < VIEW_ROTATION_DEADZONE:
-        return 0.0
-    return max(-1.0, min(1.0, axis_value))
-
-
-class DualSenseSimulator:
-    def __init__(self, controller_index: int):
-        import pygame
-
-        self.pygame: Any = pygame
-        self.controller: Any | None = None
-        self.joystick: Any | None = None
-        self.using_controller_api = False
-
-        pygame.init()
-        pygame.joystick.init()
-        count = pygame.joystick.get_count()
-        if count == 0:
-            pygame.quit()
-            raise SystemExit("No gamepad found. Use --input random to run without a controller.")
-
-        try:
-            from pygame._sdl2 import controller as sdl_controller
-
-            sdl_controller.init()
-            if sdl_controller.is_controller(controller_index):
-                self.controller = sdl_controller.Controller(controller_index)
-                self.using_controller_api = True
-        except Exception:
-            self.controller = None
-            self.using_controller_api = False
-
-        if self.controller is None:
-            self.joystick = pygame.joystick.Joystick(controller_index)
-            self.joystick.init()
-
-        self.throttle = 0.0
-        self.brake = 0.0
-        self.steering = 0.0
-        self.view_rotate_x = 0.0
-        self.view_rotate_y = 0.0
-        self.left_signal_requested = False
-        self.right_signal_requested = False
-        self.warmup_until = time.perf_counter() + GAMEPAD_WARMUP_SECONDS
-
-    def close(self) -> None:
-        if self.controller is not None:
-            self.controller.quit()
-        if self.joystick is not None:
-            self.joystick.quit()
-        self.pygame.quit()
-
-    def _button_down(self, button_indexes: tuple[int, ...]) -> bool:
-        if self.using_controller_api:
-            return self._controller_button_down(button_indexes)
-        return any(index < self.joystick.get_numbuttons() and self.joystick.get_button(index) for index in button_indexes)
-
-    def _controller_button_down(self, button_indexes: tuple[int, ...]) -> bool:
-        return any(self.controller.get_button(index) for index in button_indexes)
-
-    def _hat_left_down(self) -> bool:
-        if self.joystick is None or self.joystick.get_numhats() == 0:
-            return False
-        return self.joystick.get_hat(0)[0] < 0
-
-    def _hat_right_down(self) -> bool:
-        if self.joystick is None or self.joystick.get_numhats() == 0:
-            return False
-        return self.joystick.get_hat(0)[0] > 0
-
-    def _read_motion(self) -> tuple[float, float, float]:
-        if self.using_controller_api:
-            steering = normalize_stick(
-                normalize_signed_axis(self.controller.get_axis(self.pygame.CONTROLLER_AXIS_LEFTX))
-            )
-            brake = normalize_trigger_axis(
-                self.controller.get_axis(self.pygame.CONTROLLER_AXIS_TRIGGERLEFT)
-            )
-            throttle = normalize_trigger_axis(
-                self.controller.get_axis(self.pygame.CONTROLLER_AXIS_TRIGGERRIGHT)
-            )
-            return throttle, brake, steering
-
-        steering = normalize_stick(
-            normalize_signed_axis(self.joystick.get_axis(0) if self.joystick.get_numaxes() > 0 else 0.0)
-        )
-        brake = normalize_trigger_axis(
-            self.joystick.get_axis(4) if self.joystick.get_numaxes() > 4 else 0.0
-        )
-        throttle = normalize_trigger_axis(
-            self.joystick.get_axis(5) if self.joystick.get_numaxes() > 5 else 0.0
-        )
-        return throttle, brake, steering
-
-    def _read_view_rotation(self) -> tuple[float, float]:
-        if self.using_controller_api:
-            x_axis = normalize_view_axis(
-                normalize_signed_axis(self.controller.get_axis(self.pygame.CONTROLLER_AXIS_RIGHTX))
-            )
-            y_axis = normalize_view_axis(
-                normalize_signed_axis(self.controller.get_axis(self.pygame.CONTROLLER_AXIS_RIGHTY))
-            )
-            return x_axis, y_axis
-
-        if self.joystick is None:
-            return 0.0, 0.0
-        x_axis = normalize_view_axis(
-            normalize_signed_axis(self.joystick.get_axis(2) if self.joystick.get_numaxes() > 2 else 0.0)
-        )
-        y_axis = normalize_view_axis(
-            normalize_signed_axis(self.joystick.get_axis(3) if self.joystick.get_numaxes() > 3 else 0.0)
-        )
-        return x_axis, y_axis
-
-    def _read_signal_buttons(self) -> tuple[bool, bool]:
-        if self.using_controller_api:
-            left = any(
-                self.controller.get_button(button)
-                for button in (
-                    self.pygame.CONTROLLER_BUTTON_DPAD_LEFT,
-                    self.pygame.CONTROLLER_BUTTON_LEFTSHOULDER,
-                )
-            )
-            right = any(
-                self.controller.get_button(button)
-                for button in (
-                    self.pygame.CONTROLLER_BUTTON_DPAD_RIGHT,
-                    self.pygame.CONTROLLER_BUTTON_RIGHTSHOULDER,
-                )
-            )
-            return left, right
-
-        left = self._button_down(LEFT_SIGNAL_BUTTONS) or self._hat_left_down()
-        right = self._button_down(RIGHT_SIGNAL_BUTTONS) or self._hat_right_down()
-        return left, right
-
-    def read_input(self) -> SimulatorInput:
-        self.pygame.event.pump()
-        self.throttle, self.brake, self.steering = self._read_motion()
-        self.view_rotate_x, self.view_rotate_y = self._read_view_rotation()
-        self.left_signal_requested, self.right_signal_requested = self._read_signal_buttons()
-        if time.perf_counter() < self.warmup_until:
-            self.throttle = 0.0
-            self.brake = 0.0
-            self.steering = 0.0
-            self.view_rotate_x = 0.0
-            self.view_rotate_y = 0.0
-            self.left_signal_requested = False
-            self.right_signal_requested = False
-        return SimulatorInput(
-            throttle=self.throttle,
-            brake=self.brake,
-            steering=self.steering,
-            surround_yaw_deg=self.view_rotate_x * SURROUND_MAX_YAW_DEG,
-            surround_pitch_deg=-self.view_rotate_y * SURROUND_MAX_PITCH_DEG,
-            surround_view_active=self.view_rotate_x != 0.0 or self.view_rotate_y != 0.0,
-            left_signal_requested=self.left_signal_requested,
-            right_signal_requested=self.right_signal_requested,
-        )
-
-    def status_text(self) -> str:
-        left = "L" if self.left_signal_requested else "-"
-        right = "R" if self.right_signal_requested else "-"
-        return (
-            f"R2={self.throttle:.2f} L2={self.brake:.2f} "
-            f"LSX={self.steering:+.2f} RS={self.view_rotate_x:+.2f},{self.view_rotate_y:+.2f} "
-            f"SIG={left}{right}"
-        )
 
 
 def run_demo(
